docstring types and moved batchnorm
