+++ swyft/networks/__init__.py
@@ -1,10 +1,11 @@
-from swyft.networks.channelized import LinearWithChannel, ResidualNetWithChannel
+from swyft.networks.channelized import (
+    BatchNorm1dWithChannel,
+    LinearWithChannel,
+    ResidualNetWithChannel,
+)
 from swyft.networks.head import DefaultHead
 from swyft.networks.module import Module
-from swyft.networks.normalization import (
-    BatchNorm1dWithChannel,
-    OnlineNormalizationLayer,
-)
+from swyft.networks.normalization import OnlineNormalizationLayer
 from swyft.networks.tail import DefaultTail
 
 __all__ = [
